Fix/native android (#14)
* fix: introduces IncomingCallService

* fix: fixes errors with start service, fixes error with answer call

* fix: robust handling for observeCallInviteCancel to prevent IllegalStateException

* fix: catch crash on hangup and send it to rn

* fix: catches errors on reject

* fix: better management of vonage sdk errors

* fix: adds tryblocking on answer and adds error log

+++ android/src/main/java/com/vonagevoice/call/VonageEventsObserver.kt
@@ -87,25 +87,39 @@ class VonageEventsObserver(
             inboundCallNotifier.stopRingtoneAndInboundNotification()
             inboundCallNotifier.stopCall()
 
-            val storedCall = callRepository.getCall(callId)
-                ?: throw IllegalStateException("Call $callId does not exist on storage")
-            Log.d("VonageEventsObserver", "observeCallInviteCancel storedCall: $storedCall")
-
             val normalizedCallId = callId.lowercase()
 
-            callRepository.removeHangedUpCall(normalizedCallId)
-
-            scope.launch {
-                jsEventSender.sendCallEvent(
-                    callId = normalizedCallId,
-                    status = CallStatus.COMPLETED,
-                    phoneNumber = storedCall.phoneNumber,
-                    startedAt = storedCall.startedAt,
-                    outbound = storedCall is Call.Outbound
-                )
-            }
             notificationManager.cancelInProgressNotification()
             CallLifecycleManager.callback?.onCallEnded()
+
+            val storedCall = callRepository.getCall(callId)
+            Log.d("VonageEventsObserver", "observeCallInviteCancel storedCall: $storedCall")
+
+            if (storedCall == null) {
+                Log.e("VonageEventsObserver","Call $callId does not exist on storage")
+
+                // Fallback: storedCall is null (seen in prod) => notify RN as completed.
+                scope.launch {
+                    jsEventSender.sendCallEvent(
+                        callId = normalizedCallId,
+                        status = CallStatus.COMPLETED,
+                        phoneNumber = "",
+                        startedAt = 0.0,
+                        outbound = false
+                    )
+                }
+            } else {
+                callRepository.removeHangedUpCall(normalizedCallId)
+                scope.launch {
+                    jsEventSender.sendCallEvent(
+                        callId = normalizedCallId,
+                        status = CallStatus.COMPLETED,
+                        phoneNumber = storedCall.phoneNumber,
+                        startedAt = storedCall.startedAt,
+                        outbound = storedCall is Call.Outbound
+                    )
+                }
+            }
         }
     }
